V597: Android Studio-তে বিল্ড ভাঙা সারানো (AnatomyView Double→Float)
TK-এর স্ক্রিনশটে ১২টা 'Type mismatch: Double but Float' এসে Gradle build
ভেঙে যাচ্ছিল। আসল কারণ ksHealLump()-এ val r Double হয়ে যাওয়া, অথচ
Canvas-এর ঘরগুলো Float চায়। V589 থেকে এই ভুল ছিল।

পাহারাদারও সারানো হলো — ক্লাসপাথে আসল android.jar (API 34) যোগ,
'type mismatch' আর গোলমাল ধরা হয় না, আর নতুন কড়া নিয়ম: যে ফাইলে একটাও
unresolved reference নেই অথচ ভুল আছে, সেটা সরাসরি FAIL।

উল্টো পরীক্ষা করা হয়েছে — ভুলটা ফেরত বসালে পাহারাদার ১২টাই ধরে ফেলে।

+++ 02_ANDROID_SOURCE_CODE/PilesClinicApp/app/src/main/java/com/tkbiswas/pilesclinic/clinical/AnatomyView.kt
@@ -396,7 +396,11 @@ class AnatomyView(context: Context) : View(context) {
         val cxP = m.x + Math.cos(g.ang) * g.len * 0.50
         val cyP = m.y + Math.sin(g.ang) * g.len * 0.50
         val sc = Math.min(dst.width(), dst.height()) / 100f
-        val r = Math.max(g.len * sc * 0.62f, g.wide * sc * 0.72f)
+        // 🔴 V597 — ক্যানভাসের সব ঘর Float চায়। g.len / g.wide হলো Double, তাই
+        // গুণফলটাও Double হয়ে যাচ্ছিল — সেই কারণেই Android Studio-তে
+        // ১২টা "Type mismatch: Double but Float" ভুল আসছিল। হিসাব একই রইল,
+        // শেষে শুধু Float-এ নামানো হলো।
+        val r = Math.max(g.len * sc * 0.62f, g.wide * sc * 0.72f).toFloat()
         if (r <= 0.5f) return
         val halfPct = (r / sc).toDouble()
         val b = base
